# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type DealOfferType = typeof DealOfferType[keyof typeof DealOfferType];


export const DealOfferType = {
  'flat-rate': 'flat-rate',
  'cash-gift-card': 'cash-gift-card',
  'service-gift-card': 'service-gift-card',
  bookable: 'bookable',
} as const;

export type DealStatus = typeof DealStatus[keyof typeof DealStatus];


export const DealStatus = {
  active: 'active',
  paused: 'paused',
  archived: 'archived',
  'sold-out': 'sold-out',
  expired: 'expired',
} as const;

export interface Deal {
  id: string;
  salonName: string;
  category: string;
  city: string;
  title: string;
  offerType: DealOfferType;
  /** @nullable */
  serviceName?: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  giftCardValue?: number | null;
  image: string;
  /**
     * @minimum 0
     * @maximum 5
     */
  rating: number;
  /** @minimum 0 */
  reviewCount: number;
  /** @minimum 0 */
  originalPrice: number;
  /** @minimum 0 */
  dealPrice: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  discountPercent: number;
  /** @minimum 0 */
  savings: number;
  /** @minimum 0 */
  purchasedCount: number;
  /** @minimum 1 */
  capacity: number;
  endsAt: string;
  description: string;
  highlights: string[];
  finePrint: string;
  status: DealStatus;
  createdAt: string;
  updatedAt: string;
}

export type DealInputOfferType = typeof DealInputOfferType[keyof typeof DealInputOfferType];


export const DealInputOfferType = {
  'flat-rate': 'flat-rate',
  'cash-gift-card': 'cash-gift-card',
  'service-gift-card': 'service-gift-card',
  bookable: 'bookable',
} as const;

export interface DealInput {
  /** @minLength 1 */
  salonName: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  city: string;
  /** @minLength 1 */
  title: string;
  offerType: DealInputOfferType;
  /** @nullable */
  serviceName?: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  giftCardValue?: number | null;
  image?: string;
  /**
     * @minimum 0
     * @maximum 5
     */
  rating?: number;
  /** @minimum 0 */
  reviewCount?: number;
  /** @minimum 0 */
  originalPrice: number;
  /** @minimum 0 */
  dealPrice: number;
  /** @minimum 1 */
  capacity: number;
  endsAt: string;
  description: string;
  highlights: string[];
  finePrint: string;
}

export type DealUpdateOfferType = typeof DealUpdateOfferType[keyof typeof DealUpdateOfferType];


export const DealUpdateOfferType = {
  'flat-rate': 'flat-rate',
  'cash-gift-card': 'cash-gift-card',
  'service-gift-card': 'service-gift-card',
  bookable: 'bookable',
} as const;

export type DealUpdateStatus = typeof DealUpdateStatus[keyof typeof DealUpdateStatus];


export const DealUpdateStatus = {
  active: 'active',
  paused: 'paused',
  archived: 'archived',
  'sold-out': 'sold-out',
  expired: 'expired',
} as const;

export interface DealUpdate {
  /** @minLength 1 */
  salonName?: string;
  /** @minLength 1 */
  category?: string;
  /** @minLength 1 */
  city?: string;
  /** @minLength 1 */
  title?: string;
  offerType?: DealUpdateOfferType;
  /** @nullable */
  serviceName?: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  giftCardValue?: number | null;
  image?: string;
  /**
     * @minimum 0
     * @maximum 5
     */
  rating?: number;
  /** @minimum 0 */
  reviewCount?: number;
  /** @minimum 0 */
  originalPrice?: number;
  /** @minimum 0 */
  dealPrice?: number;
  /** @minimum 1 */
  capacity?: number;
  endsAt?: string;
  description?: string;
  highlights?: string[];
  finePrint?: string;
  status?: DealUpdateStatus;
}

export interface FavoriteIds {
  favoriteDealIds: string[];
}

export interface PurchaseInput {
  /**
     * @minimum 1
     * @maximum 4
     */
  quantity: number;
}

export type VoucherStatus = typeof VoucherStatus[keyof typeof VoucherStatus];


export const VoucherStatus = {
  active: 'active',
  used: 'used',
  expired: 'expired',
} as const;

export interface Voucher {
  id: string;
  dealId: string;
  dealTitle: string;
  salonName: string;
  customerName: string;
  purchaseDate: string;
  expiresAt: string;
  code: string;
  status: VoucherStatus;
  qrSeed: string;
  /** @nullable */
  redeemedAt?: string | null;
}

export interface PurchaseResponse {
  vouchers: Voucher[];
}

/**
 * Invalid request
 */
export type BadRequestResponse = Error;

/**
 * Authentication required
 */
export type UnauthorizedResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;