# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: deals
    description: Marketplace and salon deal operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /deals:
    get:
      operationId: listPublicDeals
      tags: [deals]
      summary: List active marketplace deals
      responses:
        "200":
          description: Active deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
  /deals/{id}/purchase:
    post:
      operationId: purchaseDeal
      tags: [deals]
      summary: Purchase one or more vouchers for a deal
      parameters:
        - $ref: "#/components/parameters/DealId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "201":
          description: Purchased vouchers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /me/favorites:
    get:
      operationId: listFavoriteIds
      tags: [favorites]
      summary: List the signed-in customer's favorite deals
      responses:
        "200":
          description: Favorite deal IDs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteIds"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/favorites/{id}:
    put:
      operationId: addFavorite
      tags: [favorites]
      summary: Save a deal to the signed-in customer's favorites
      parameters:
        - $ref: "#/components/parameters/DealId"
      responses:
        "200":
          description: Updated favorite deal IDs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteIds"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove a deal from the signed-in customer's favorites
      parameters:
        - $ref: "#/components/parameters/DealId"
      responses:
        "200":
          description: Updated favorite deal IDs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteIds"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /me/vouchers:
    get:
      operationId: listMyVouchers
      tags: [vouchers]
      summary: List the signed-in customer's vouchers
      responses:
        "200":
          description: Customer vouchers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Voucher"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /salon/vouchers:
    get:
      operationId: listSalonVouchers
      tags: [vouchers]
      summary: List vouchers for the signed-in salon
      responses:
        "200":
          description: Salon vouchers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Voucher"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /salon/vouchers/{code}/redeem:
    post:
      operationId: redeemSalonVoucher
      tags: [vouchers]
      summary: Redeem a voucher owned by the signed-in salon
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Redeemed voucher
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Voucher"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /salon/deals:
    get:
      operationId: listSalonDeals
      tags: [deals]
      summary: List deals owned by the signed-in salon
      responses:
        "200":
          description: The salon owner's deals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deal"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createSalonDeal
      tags: [deals]
      summary: Create a deal for the signed-in salon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealInput"
      responses:
        "201":
          description: Created deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /salon/deals/{id}:
    patch:
      operationId: updateSalonDeal
      tags: [deals]
      summary: Update a deal owned by the signed-in salon
      parameters:
        - $ref: "#/components/parameters/DealId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealUpdate"
      responses:
        "200":
          description: Updated deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    DealId:
      name: id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Deal:
      type: object
      required:
        - id
        - salonName
        - category
        - city
        - title
        - offerType
        - image
        - rating
        - reviewCount
        - originalPrice
        - dealPrice
        - discountPercent
        - savings
        - purchasedCount
        - capacity
        - endsAt
        - description
        - highlights
        - finePrint
        - status
        - createdAt
        - updatedAt
      properties:
        id: { type: string }
        salonName: { type: string }
        category: { type: string }
        city: { type: string }
        title: { type: string }
        offerType:
          type: string
          enum: [flat-rate, cash-gift-card, service-gift-card, bookable]
        serviceName: { type: ["string", "null"] }
        giftCardValue: { type: ["integer", "null"], minimum: 0 }
        image: { type: string }
        rating: { type: number, minimum: 0, maximum: 5 }
        reviewCount: { type: integer, minimum: 0 }
        originalPrice: { type: integer, minimum: 0 }
        dealPrice: { type: integer, minimum: 0 }
        discountPercent: { type: integer, minimum: 0, maximum: 100 }
        savings: { type: integer, minimum: 0 }
        purchasedCount: { type: integer, minimum: 0 }
        capacity: { type: integer, minimum: 1 }
        endsAt: { type: string, format: date }
        description: { type: string }
        highlights:
          type: array
          items: { type: string }
        finePrint: { type: string }
        status:
          type: string
          enum: [active, paused, archived, sold-out, expired]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DealInput:
      type: object
      required:
        - salonName
        - category
        - city
        - title
        - offerType
        - originalPrice
        - dealPrice
        - capacity
        - endsAt
        - description
        - highlights
        - finePrint
      properties:
        salonName: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        city: { type: string, minLength: 1 }
        title: { type: string, minLength: 1 }
        offerType:
          type: string
          enum: [flat-rate, cash-gift-card, service-gift-card, bookable]
        serviceName: { type: ["string", "null"] }
        giftCardValue: { type: ["integer", "null"], minimum: 0 }
        image: { type: string }
        rating: { type: number, minimum: 0, maximum: 5 }
        reviewCount: { type: integer, minimum: 0 }
        originalPrice: { type: integer, minimum: 0 }
        dealPrice: { type: integer, minimum: 0 }
        capacity: { type: integer, minimum: 1 }
        endsAt: { type: string, format: date }
        description: { type: string }
        highlights:
          type: array
          items: { type: string }
        finePrint: { type: string }
    DealUpdate:
      type: object
      properties:
        salonName: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        city: { type: string, minLength: 1 }
        title: { type: string, minLength: 1 }
        offerType:
          type: string
          enum: [flat-rate, cash-gift-card, service-gift-card, bookable]
        serviceName: { type: ["string", "null"] }
        giftCardValue: { type: ["integer", "null"], minimum: 0 }
        image: { type: string }
        rating: { type: number, minimum: 0, maximum: 5 }
        reviewCount: { type: integer, minimum: 0 }
        originalPrice: { type: integer, minimum: 0 }
        dealPrice: { type: integer, minimum: 0 }
        capacity: { type: integer, minimum: 1 }
        endsAt: { type: string, format: date }
        description: { type: string }
        highlights:
          type: array
          items: { type: string }
        finePrint: { type: string }
        status:
          type: string
          enum: [active, paused, archived, sold-out, expired]
    FavoriteIds:
      type: object
      required: [favoriteDealIds]
      properties:
        favoriteDealIds:
          type: array
          items: { type: string }
    PurchaseInput:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer, minimum: 1, maximum: 4 }
    PurchaseResponse:
      type: object
      required: [vouchers]
      properties:
        vouchers:
          type: array
          items:
            $ref: "#/components/schemas/Voucher"
    Voucher:
      type: object
      required:
        - id
        - dealId
        - dealTitle
        - salonName
        - customerName
        - purchaseDate
        - expiresAt
        - code
        - status
        - qrSeed
      properties:
        id: { type: string }
        dealId: { type: string }
        dealTitle: { type: string }
        salonName: { type: string }
        customerName: { type: string }
        purchaseDate: { type: string, format: date }
        expiresAt: { type: string, format: date }
        code: { type: string }
        status:
          type: string
          enum: [active, used, expired]
        qrSeed: { type: string }
        redeemedAt: { type: ["string", "null"], format: date-time }