# lib/api-zod/src/generated/types/deal.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DealOfferType } from './dealOfferType';
import type { DealStatus } from './dealStatus';

export interface Deal {
  id: string;
  salonName: string;
  category: string;
  city: string;
  title: string;
  offerType: DealOfferType;
  /** @nullable */
  serviceName?: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  giftCardValue?: number | null;
  image: string;
  /**
     * @minimum 0
     * @maximum 5
     */
  rating: number;
  /** @minimum 0 */
  reviewCount: number;
  /** @minimum 0 */
  originalPrice: number;
  /** @minimum 0 */
  dealPrice: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  discountPercent: number;
  /** @minimum 0 */
  savings: number;
  /** @minimum 0 */
  purchasedCount: number;
  /** @minimum 1 */
  capacity: number;
  endsAt: Date;
  description: string;
  highlights: string[];
  finePrint: string;
  status: DealStatus;
  createdAt: Date;
  updatedAt: Date;
}