# lib/api-zod/src/generated/types/dealUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DealUpdateOfferType } from './dealUpdateOfferType';
import type { DealUpdateStatus } from './dealUpdateStatus';

export interface DealUpdate {
  /** @minLength 1 */
  salonName?: string;
  /** @minLength 1 */
  category?: string;
  /** @minLength 1 */
  city?: string;
  /** @minLength 1 */
  title?: string;
  offerType?: DealUpdateOfferType;
  /** @nullable */
  serviceName?: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  giftCardValue?: number | null;
  image?: string;
  /**
     * @minimum 0
     * @maximum 5
     */
  rating?: number;
  /** @minimum 0 */
  reviewCount?: number;
  /** @minimum 0 */
  originalPrice?: number;
  /** @minimum 0 */
  dealPrice?: number;
  /** @minimum 1 */
  capacity?: number;
  endsAt?: Date;
  description?: string;
  highlights?: string[];
  finePrint?: string;
  status?: DealUpdateStatus;
}