# lib/api-zod/src/generated/types/purchaseInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PurchaseInput {
  /**
     * @minimum 1
     * @maximum 4
     */
  quantity: number;
}