# lib/api-zod/src/generated/types/purchaseResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Voucher } from './voucher';

export interface PurchaseResponse {
  vouchers: Voucher[];
}