# lib/api-zod/src/generated/types/voucher.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VoucherStatus } from './voucherStatus';

export interface Voucher {
  id: string;
  dealId: string;
  dealTitle: string;
  salonName: string;
  customerName: string;
  purchaseDate: Date;
  expiresAt: Date;
  code: string;
  status: VoucherStatus;
  qrSeed: string;
  /** @nullable */
  redeemedAt?: Date | null;
}